+++ cmass/diagnostics/summ.py
@@ -17,91 +17,8 @@ from ..nbody.tools import parse_nbody_config
 from ..bias.apply_hod import parse_hod
 from .tools import MA, MAz, get_box_catalogue, get_box_catalogue_rsd
 from .tools import calcPk, calcBk_bfast, get_mesh_resolution
+from .tools import store_summary
 from ..survey.tools import sky_to_xyz
-
-
-def get_binning(summary, L, N, threads, rsd=False):
-    ells = [0,] if not rsd else [0, 2, 4]
-    if summary == 'Pk':
-        return {
-            'k_edges': np.linspace(0, 1., 31),
-            'n_mesh': N,
-            'los': 'z',
-            'compensations': 'ngp',
-            'ells': ells,
-        }
-    if summary == 'Bk':
-        k_min = 1.05*2 * np.pi / L
-        n_mesh = 64
-        k_max = 0.95 * np.pi * n_mesh / L
-        num_bins = 15
-        return {
-            'k_bins': np.logspace(np.log10(k_min), np.log10(k_max), num_bins),
-            'n_mesh': n_mesh,
-            'lmax': 2,
-            'ells': ells,
-        }
-    if summary == 'TwoPCF':
-        num_bins = 60
-        return {
-            'r_bins': np.logspace(-2, np.log10(150.), num_bins),
-            'mu_bins': np.linspace(-1., 1., 201),
-            'ells': ells,
-            'n_threads': threads,
-            'los': 'z',
-        }
-    if summary == 'WST':
-        num_bins = 60
-        return {
-            'J_3d': 3,
-            'L_3d': 3,
-            'integral_powers': [0.8,],
-            'sigma': 0.8,
-            'n_mesh': N,
-        }
-    if summary == 'DensitySplit':
-        num_bins = 60
-        return {
-            'r_bins': np.logspace(-1, np.log10(150.), num_bins),
-            'mu_bins': np.linspace(-1., 1., 201),
-            'n_quantiles': 5,
-            'smoothing_radius': 10.0,
-            'ells': ells,
-            'n_threads': threads,
-        }
-    if summary == 'KNN':
-        num_bins = 60
-        return {
-            'r_bins': np.logspace(-2, np.log10(30.), num_bins),
-            'k': [1, 3, 7, 11],
-            'n_threads': threads,
-        }
-    else:
-        raise NotImplementedError(f'{summary} not implemented')
-
-
-def store_summary(
-    catalog, group, summary_name,
-    box_size, num_bins, num_threads, use_rsd=False
-):
-    # get summary binning
-    binning_config = get_binning(
-        summary_name, box_size, num_bins, num_threads, rsd=use_rsd)
-
-    logging.info(f'Computing Summary: {summary_name}')
-
-    # compute summary
-    import summarizer  # only import if needed. TODO: get working
-    summary_function = getattr(summarizer, summary_name)(**binning_config)
-    summary_data = summary_function(catalog)
-
-    # store summary
-    summary_dataset = summary_function.to_dataset(summary_data)
-    for coord_name, coord_value in summary_dataset.coords.items():
-        dataset_key = f"{'z' if use_rsd else ''}{summary_name}_{coord_name}"
-        group.create_dataset(dataset_key, data=coord_value.values)
-    summary_key = summary_name if not use_rsd else f'z{summary_name}'
-    group.create_dataset(summary_key, data=summary_dataset.values)
 
 
 def run_pylians(
